Hopefully works in MP now..
Lots of bugs found in MP (as usual).

The trigger list is not synchronised between server/client, and it seems too volatile to attempt this.  Instead now uses object ID to find the appropriate trigger, not the index.

Open/close cover on multiple fill type tools must also be network synchronised.  Added a whole function, but not sure why the original game function was not doing this already..

Ready for more testing.